Fixed here_doc struct with expansion flag
