+++ includes/parser_prototypes.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   parser_prototypes.h                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: harsh <harsh@student.42.fr>                +#+  +:+       +#+        */
+/*   By: hkumbhan <hkumbhan@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/12 14:47:43 by cwenz             #+#    #+#             */
-/*   Updated: 2023/11/20 22:39:05 by harsh            ###   ########.fr       */
+/*   Updated: 2023/11/21 17:36:26 by hkumbhan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 
 /* ---------------------------------- Lexer --------------------------------- */
 
-t_lexer	*tokenise(char *input);
+t_lexer	*tokenise(char *input, t_minishell *minishell);
 int		token_word(t_lexer **head, char *str);
 int		token_pipe(t_lexer **head, char *str);
 int		token_squote(t_lexer **head, char *str);
@@ -31,7 +31,9 @@ int		create_dbgreater(t_lexer **head, char *str);
 /* --------------------------------- Parser --------------------------------- */
 
 int		parser(t_lexer **lexer, t_minishell *minishell);
+void	put_args(t_cmd **cmd_table, t_lexer **lexer, t_minishell *minishell);
 int		count_pipes(t_lexer **lexer);
+int		t_check(t_token token_type);
 void	init_t_cmd(t_cmd **cmd);
 void	free_cmd_table(t_cmd **cmd_table);
 int		count_words(t_lexer **lexer);
@@ -40,5 +42,12 @@ void	print_cmd_table(t_cmd **cmd_table);
 int		syntax_checker(t_lexer **lexer, t_minishell *minishell);
 void	redir_add_back(t_redir	**head, t_redir *new_node);
 void	hd_add_back(t_heredoc **head, t_heredoc *new);
+void	move_and_free(t_lexer **lexer);
+
+/* EXPANDER */
+
+int	check_var(char c);
+int	special_var(char **ret, t_minishell *minishell);
+int	special_char(char **ret, char *str);
 
 #endif
