+++ src/expander/expander.c
@@ -1,34 +1,6 @@
 #include "minishell.h"
 
-int	check_var(char c)
-{
-	if (ft_isalnum(c) == TRUE || c == '_')
-		return (TRUE);
-	return (FALSE);
-}
-
-int	special_var(char **ret, t_minishell *minishell)
-{
-	char	*tmp;
-
-	tmp = ft_itoa(minishell->exit_code);
-	*ret = ft_strjoin(*ret, tmp);
-	free(tmp);
-	return (2);
-}
-
-int	special_char(char **ret, char *str)
-{
-	char	*tmp;
-
-	tmp = NULL;
-	tmp = ft_strndup(str, 2);
-	*ret = ft_strjoin_gnl(*ret, tmp);
-	free(tmp);
-	return (2);
-}
-
-int	get_var(char **ret, char *str, t_minishell *minshell)
+static int	get_var(char **ret, char *str, t_minishell *minshell)
 {
 	int		i;
 	t_env	*env;
@@ -52,16 +24,28 @@ int	get_var(char **ret, char *str, t_minishell *minshell)
 	return (i + 1);
 }
 
+static void	ft_join_rest(int *i, t_lexer *lexer, char **ret)
+{
+	char	*tmp;
+	int		j;
+
+	tmp = NULL;
+	j = 0;
+	while ((*i) + j < lexer->strlen && lexer->start[*i + j] != '$')
+		j++;
+	tmp = ft_strndup(&lexer->start[*i], j);
+	(*ret) = ft_strjoin_gnl(*ret, tmp);
+	free(tmp);
+	(*i) += j;
+}
+
 char	*expander(t_lexer *lexer, t_minishell *minishell)
 {
 	char	*ret;
-	 char	*tmp;
 	int		i;
-	int		j;
 
 	i = 0;
 	ret = NULL;
-	tmp = NULL;
 	while (i < lexer->strlen && lexer->start[i] != '$')
 		i++ ;
 	if (i == lexer->strlen)
@@ -78,16 +62,8 @@ char	*expander(t_lexer *lexer, t_minishell *minishell)
 		else if (lexer->start[i] == '$' && check_var(lexer->start[i + 1]) == TRUE)
 			i += get_var(&ret, &lexer->start[i + 1], minishell);
 		else
-		{
-			j = 0;
-			while (i + j < lexer->strlen && lexer->start[i + j] != '$')
-				j++;
-			tmp = ft_strndup(&lexer->start[i], j);
-			ret = ft_strjoin_gnl(ret, tmp);
-			free(tmp);
-			i += j;
-		}
+			ft_join_rest(&i, lexer, &ret);
 	}
-	//printf("Expanded string: %s\n", ret);
+	printf("Expanded string: %s\n", ret);
 	return (ret);
 }
