+++ src/lexer/lexer.c
@@ -6,7 +6,7 @@
 /*   By: hkumbhan <hkumbhan@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/17 10:26:47 by hkumbhan          #+#    #+#             */
-/*   Updated: 2023/11/07 12:37:54 by hkumbhan         ###   ########.fr       */
+/*   Updated: 2023/11/21 17:36:09 by hkumbhan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 /// @brief Checks for open Quotes, Escape characters \ and ;
 /// @param input Input string read from terminal
 /// @return FALSE if open quotes or esc char found else TRUE
-int	check_input(char *input)
+int	check_input(char *input, t_minishell *minishell)
 {
 	char	c;
 
@@ -31,6 +31,7 @@ int	check_input(char *input)
 	if (c == '\0')
 		return (TRUE);
 	ft_putendl_fd(QUOTES, STDERR_FILENO);
+	minishell->exit_code = EXIT_FAILURE;
 	return (FALSE);
 }
 
@@ -57,14 +58,14 @@ int	token_redirect(t_lexer **head, char *str)
 /// @brief Creates tokens and add them to a list
 /// @param input 
 /// @return 
-t_lexer	*tokenise(char *input)
+t_lexer	*tokenise(char *input, t_minishell *minishell)
 {
 	t_lexer	*head;
 	int		i;
 
 	i = 0;
 	head = NULL;
-	if (check_input(input) == FALSE)
+	if (check_input(input, minishell) == FALSE)
 		return (NULL);
 	while (input[i] != '\0')
 	{
