+++ src/parser/parser_utils2.c
@@ -40,3 +40,39 @@ void	hd_add_back(t_heredoc **head, t_heredoc *new)
 	}
 	return ;
 }
+
+void	print_files(t_redir *head)
+{
+	while(head != NULL)
+	{
+		printf("File name: %s\n", head->file_name);
+		printf("File type: %d\n", head->file_type);
+		head = head->next;
+	}
+}
+
+void	print_cmd_table(t_cmd **cmd_table)
+{
+	t_cmd		*tmp;
+	t_redir		*redir;
+	t_heredoc	*hd;
+	int			k;
+
+	k = -1;
+	tmp = (*cmd_table);
+	redir = tmp->files;
+	hd = tmp->heredoc;
+	printf("\nCommand table: \n");
+	if (tmp != NULL)
+	{
+		while (tmp->cmd[++k] != NULL)
+			printf("cmd[%d]: %s\n", k, tmp->cmd[k]);
+	}
+	while (hd != NULL)
+	{
+		printf("heredoc: %s\n", hd->str);
+		printf("Expand: %s\n", hd->expand ? "TRUE" : "FALSE");
+		hd = hd->next;
+	}
+	print_files(redir);
+}
