+++ src/parser/syntax_error.c
@@ -10,7 +10,7 @@ int	ft_error(t_minishell *minishell, t_token token)
 
 	ft_putstr_fd(SYNTAX, STDERR_FILENO);
 	ft_putendl_fd(token_str[token], STDERR_FILENO);
-	minishell->exit_code = 258;
+	minishell->exit_code = 3;
 	return (FAIL);
 }
